pyproject.toml
[project]
name = "strava-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.27.0",
    "openai-agents>=0.14.7",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
